requirements.txt
annotated-types==0.6.0
anyio==3.7.1
certifi==2023.11.17
distro==1.8.0
exceptiongroup==1.2.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.4
openai==1.3.5
pydantic==2.5.2
pydantic_core==2.14.5
PyQt5==5.15.10
PyQt5-Qt5==5.15.11
PyQt5-sip==12.13.0
python-dotenv==1.0.0
sniffio==1.3.0
tqdm==4.66.1
typing_extensions==4.8.0
